Interface

To extend the above functionality to a new array type, you should use the types and implement the interfaces listed on this page. GPUArrays is design around having two different array types to represent a GPU array: one that only ever lives on the host, and one that actually can be instantiated on the device (i.e. in kernels).

Host-side

Your host-side array type should build on the AbstractGPUArray supertype:

GPUArrays.AbstractGPUArray — Type

AbstractGPUArray{T, N} <: DenseArray{T, N}

Supertype for N-dimensional GPU arrays (or array-like types) with elements of type T. Instances of this type are expected to live on the host, see AbstractDeviceArray for device-side objects.

First of all, you should implement operations that are expected to be defined for any AbstractArray type. Refer to the Julia manual for more details, or look at the JLArray reference implementation.

To be able to actually use the functionality that is defined for AbstractGPUArrays, you should provide implementations of the following interfaces:

GPUArrays.unsafe_reinterpret — Function

unsafe_reinterpret(T, a, dims)

Reinterpret the array a to have a new element type T and size dims.

Devices

GPUArrays.device — Function

device(A::AbstractArray)

Gets the device associated to the Array A

GPUArrays.synchronize — Function

synchronize(A::AbstractArray)

Blocks until all operations are finished on A

Execution

Missing docstring.

Missing docstring for GPUArrays.AbstractGPUBackend. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.backend. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays._gpu_call. Check Documenter's build log for details.

Linear algebra

Missing docstring.

Missing docstring for GPUArrays.blas_module. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.blasbuffer. Check Documenter's build log for details.

Device-side

To work with GPU memory on the device itself, e.g. within a kernel, we need a different type: Most functionality will behave differently when running on the GPU, e.g., accessing memory directly instead of copying it to the host. We should also take care not to call into any host library, such as the Julia runtime or the system's math library.

GPUArrays.AbstractDeviceArray — Type

AbstractDeviceArray{T, N} <: DenseArray{T, N}

Supertype for N-dimensional GPU arrays (or array-like types) with elements of type T. Instances of this type are expected to live on the device, see AbstractGPUArray for device-side objects.

Your device array type should again implement the core elements of the AbstractArray interface, such as indexing and certain getters. Refer to the Julia manual for more details, or look at the JLDeviceArray reference implementation.

You should also provide implementations of several "GPU intrinsics". To make sure the correct implementation is called, the first argument to these intrinsics will be the kernel state object from before.

GPUArrays.LocalMemory — Function

Creates a block local array pointer with T being the element type and N the length. Both T and N need to be static! C is a counter for approriately get the correct Local mem id in CUDAnative. This is an internal method which needs to be overloaded by the GPU Array backends

GPUArrays.synchronize_threads — Function

 synchronize_threads(state)

in CUDA terms __synchronize in OpenCL terms: barrier(CLK_LOCAL_MEM_FENCE)

Missing docstring.

Missing docstring for GPUArrays.blockidx_x. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.blockidx_y. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.blockidx_z. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.blockdim_x. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.blockdim_y. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.blockdim_z. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.threadidx_x. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.threadidx_y. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.threadidx_z. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.griddim_x. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.griddim_y. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GPUArrays.griddim_z. Check Documenter's build log for details.